+++ tket/include/tket/Gate/GateUnitaryMatrixImplementations.hpp
@@ -112,6 +112,8 @@ struct GateUnitaryMatrixImplementations {
   static Eigen::Matrix2cd PhasedX(double alpha, double beta);
   static Eigen::MatrixXcd NPhasedX(
       unsigned int number_of_qubits, double alpha, double beta);
+  static Eigen::Matrix4cd TwinPhasedX(double alpha, double beta);
+  static Eigen::Matrix4cd PhasedXX(double alpha, double beta);
 
   // Because number_of_qubits is unrestricted, it might also be useful
   // to have a sparse version.